addition tool
This is a simple addition and multiplication tool written as a college exercise. I might need to update, reuse or adapt for use on something else later but for the time being I am saving it here. It utilises filter commands to extract usable data from user input and a repeat loop at the end.