Tax calculator
This is a flat tax calculator tool written as a college exercise. Unlike earlier more simplistic project this one uses the filter command to extract usable data from user input as well as a repeat loop at the end to allow repetition if necessary. I might need to update, reuse or adapt for something else later but for the time being I am saving it here.